+++ R/derive_locf_records.R
@@ -1,7 +1,8 @@
 #' Derive LOCF (Last Observation Carried Forward) Records
 #'
 #' Adds LOCF records as new observations for each 'by group' when the dataset
-#' does not contain observations for missed visits/time points.
+#' does not contain observations for missed visits/time points and when analysis
+#' value is missing.
 #'
 #' @param dataset
 #' `r roxygen_param_dataset(expected_vars = c("by_vars", "analysis_var", "order", "keep_vars"))`
@@ -17,7 +18,7 @@
 #'
 #'   For each group defined by `by_vars` those observations from `dataset_ref`
 #'   are added to the output dataset which do not have a corresponding observation
-#'   in the input dataset or for which `analysis_var` is `NA` for the corresponding observation
+#'   in the input dataset or for which `analysis_var` is NA for the corresponding observation
 #'   in the input dataset.
 #'
 #'   `r roxygen_param_by_vars()`
@@ -28,22 +29,30 @@
 #'  The variables to group by in `dataset_ref` when determining which observations should be added
 #'  to the input dataset.
 #'
-#'   *Default*: All the variables in `dataset_ref`
-#'
 #'   `r roxygen_param_by_vars()`
+#'
+#' @default All the variables in `dataset_ref`
 #'
 #'
 #' @param analysis_var Analysis variable.
 #'
-#'   *Default*: `AVAL`
-#'
 #' @permitted a variable
 #'
-#' @param analysis_var_fill Update `analysis_var` when `NA` without adding new observations?
+#' @param imputation Select the mode of imputation:
 #'
-#'   *Default*: `FALSE`
+#'   `add`: Keep all original records and add imputed records for missing
+#'   timepoints (from `dataset_ref`) and missing `analysis_var` values, populating
+#'   only `by_vars`, `order`, and `id_vars_ref` in the new records.
 #'
-#'   *Permitted Values:* `TRUE`, `FALSE`
+#'   `update`: Update records with missing `analysis_var`, and add imputed records
+#'   for missing timepoints (from `dataset_ref`).
+#'
+#'   `update_add`: Keep all original records and add imputed records for missing
+#'   timepoints (from `dataset_ref`) and missing `analysis_var`, populating all
+#'   other variables in the new records.
+#'
+#'
+#' @permitted One of these 3 values: `"add"`, `"update"`, `"update_add"`
 #'
 #' @param order Sort order
 #'
@@ -92,20 +101,11 @@
 #'   "CDISC01", "01-701-1015", "DIABP",        2,    NA,        4, "WEEK 4",
 #'   "CDISC01", "01-701-1015", "DIABP",        2,    NA,        6, "WEEK 6",
 #'   "CDISC01", "01-701-1015", "SYSBP",        3,   130,        0, "BASELINE",
-#'   "CDISC01", "01-701-1015", "SYSBP",        3,   132,        2, "WEEK 2",
-#'   "CDISC01", "01-701-1028", "PULSE",        1,    61,        0, "BASELINE",
-#'   "CDISC01", "01-701-1028", "PULSE",        1,    60,        6, "WEEK 6",
-#'   "CDISC01", "01-701-1028", "DIABP",        2,    51,        0, "BASELINE",
-#'   "CDISC01", "01-701-1028", "DIABP",        2,    50,        2, "WEEK 2",
-#'   "CDISC01", "01-701-1028", "DIABP",        2,    51,        4, "WEEK 4",
-#'   "CDISC01", "01-701-1028", "DIABP",        2,    50,        6, "WEEK 6",
-#'   "CDISC01", "01-701-1028", "SYSBP",        3,   121,        0, "BASELINE",
-#'   "CDISC01", "01-701-1028", "SYSBP",        3,   121,        2, "WEEK 2",
-#'   "CDISC01", "01-701-1028", "SYSBP",        3,   121,        4, "WEEK 4",
-#'   "CDISC01", "01-701-1028", "SYSBP",        3,   121,        6, "WEEK 6"
+#'   "CDISC01", "01-701-1015", "SYSBP",        3,   132,        2, "WEEK 2"
 #' )
 #'
-#' # A dataset with all the combinations of PARAMCD, PARAM, AVISIT, AVISITN, ... which are expected.
+#' # A dataset with all the combinations of PARAMCD, PARAM, AVISIT, AVISITN, ...
+#' # which are expected.
 #' advs_expected_obsv <- tribble(
 #'   ~PARAMCD, ~AVISITN, ~AVISIT,
 #'   "PULSE",         0, "BASELINE",
@@ -124,6 +124,8 @@
 #'   dataset = advs,
 #'   dataset_ref = advs_expected_obsv,
 #'   by_vars = exprs(STUDYID, USUBJID, PARAMCD),
+#'   analysis_var = AVAL,
+#'   imputation = "update_add",
 #'   order = exprs(AVISITN, AVISIT),
 #'   keep_vars = exprs(PARAMN)
 #' ) |>
@@ -134,7 +136,7 @@ derive_locf_records <- function(dataset,
                                 by_vars,
                                 id_vars_ref = lapply(names(dataset_ref), sym),
                                 analysis_var = AVAL,
-                                analysis_var_fill = FALSE,
+                                imputation = "add",
                                 order,
                                 keep_vars = NULL) {
   #### Input Checking ####
@@ -145,7 +147,13 @@ derive_locf_records <- function(dataset,
   assert_vars(keep_vars, optional = TRUE)
   assert_vars(id_vars_ref, optional = TRUE)
   assert_expr_list(order)
-  assert_logical_scalar(analysis_var_fill)
+
+  imputation <-
+    assert_character_scalar(
+      imputation,
+      values = c("add", "update", "update_add"),
+      case_sensitive = FALSE
+    )
 
   # Check by_vars and order variables in input datasets
   assert_data_frame(dataset_ref)
@@ -160,9 +168,7 @@ derive_locf_records <- function(dataset,
 
   #### Prepping 'dataset_ref' ####
 
-
   # Get the IDs from input dataset for which the expected observations are to be added
-
   ids <- dataset %>%
     select(!!!setdiff(by_vars, chr2vars(colnames(dataset_ref)))) %>%
     distinct()
@@ -174,71 +180,89 @@ derive_locf_records <- function(dataset,
 
   ##### Add LOCF records ####
 
-  # Split input dataset into the missing and non-missing analysis_var (e.g., AVAL) records
-  aval_missing <- dataset %>%
-    filter(is.na(!!analysis_var))
-
-  aval_not_missing <- dataset %>%
-    drop_na(!!analysis_var)
-
 
   # Get the variable names from the expected observation dataset
   exp_obs_vars <- exp_obsv %>%
     colnames()
 
-  # Get the variable names from the expected observation dataset to join by
-  exp_obs_by_vars <- exp_obsv %>%
-    select(!!!by_vars, !!!id_vars_ref) %>%
-    colnames()
+  # Get the variable names to join by
+  exp_obs_by_vars <- as.character(union(by_vars, id_vars_ref))
 
 
   # Get unique combination of visits/timepoints per parameter per subject
-  # from the original input dataset (with non-missing analysis_var)
-  advs_unique_original <- aval_not_missing %>%
+  # from the input dataset
+  advs_unique_original <- dataset %>%
     select(all_of(exp_obs_by_vars)) %>%
     distinct()
 
 
   tmp_dtype <- get_new_tmp_var(exp_obsv, prefix = "tmp_dtype")
 
-  # Get all the expected observations that are to be added to the input
-  # dataset (with non-missing analysis_var)
-  advs_exp_obsv3 <- exp_obsv %>%
-    mutate(!!tmp_dtype := "LOCF") %>%
+  # Get all the expected observations that are to be added to the input dataset
+  advs_exp_obsv <- exp_obsv %>%
     anti_join(advs_unique_original, by = c(exp_obs_by_vars))
 
-  # Merge the expected observations with the input dataset (with non-missing analysis_var)
+
+  # Add the expected observations to the input dataset
   # Arrange the dataset by 'order' and group it by 'by_vars'
-  # Use fill() to fill the analysis_var from the previous observation for the newly added records
+  # Use fill() to fill the 'analysis_var' from the previous observation for the newly
+  # added records
 
+  aval_locf <- dataset %>%
+    full_join(advs_exp_obsv, by = c(exp_obs_vars)) %>%
+    mutate(!!tmp_dtype := if_else(is.na(!!analysis_var), "LOCF", NA_character_))
 
-  aval_not_missing_locf <- aval_not_missing %>%
-    full_join(advs_exp_obsv3, by = c(exp_obs_vars))
-
-  if ("DTYPE" %in% colnames(aval_not_missing)) {
-    aval_not_missing_locf <- aval_not_missing_locf %>%
+  if ("DTYPE" %in% colnames(aval_locf)) {
+    aval_locf <- aval_locf %>%
       mutate(DTYPE = if_else(!!tmp_dtype == "LOCF", "LOCF", DTYPE, missing = DTYPE)) %>%
       select(-!!tmp_dtype)
   } else {
-    aval_not_missing_locf <- rename(aval_not_missing_locf, DTYPE = !!tmp_dtype)
+    aval_locf <- rename(aval_locf, DTYPE = !!tmp_dtype)
   }
 
-  aval_not_missing_locf <- aval_not_missing_locf %>%
+  aval_locf <- aval_locf %>%
     arrange(!!!by_vars, !!!order) %>%
     group_by(!!!by_vars) %>%
     fill(!!analysis_var, !!!keep_vars) %>%
     ungroup()
 
 
+  # When imputation = 'add',  keep all variables other than analysis_var, by_vars,
+  # order, id_vars_ref, keep_vars and 'DTYPE' missing. Else, keep all variables populated
+
+  if (imputation == "add") {
+    # Non-imputed records
+    non_locf <- aval_locf %>%
+      filter(!(DTYPE %in% c("LOCF")))
+
+    # imputed records
+    locf <- aval_locf %>%
+      filter(DTYPE %in% c("LOCF"))
+
+    aval_locf <- locf %>%
+      mutate(across(
+        .cols = -as.character(c(c(analysis_var, by_vars, order, id_vars_ref, keep_vars), "DTYPE")),
+        .fns = ~ vector(typeof(.), 1)[NA]
+      )) %>%
+      bind_rows(non_locf)
+  } else if (imputation %in% c("update", "update_add")) {
+    aval_locf <- aval_locf
+  }
+
+
 
   # Output dataset:
-  # If analysis_var_fill == FALSE, merge the analysis_var missing with non-missing+newly added LOCF
-  # records
-  # If analysis_var_fill == TRUE, keep non-missing+newly added LOCF records
+  # If imputation == 'add' or 'update_add', add the missing analysis_var records
+  # with non-missing + newly added LOCF records
+  # If imputation == 'update', keep non-missing + newly added LOCF records
 
-  if (analysis_var_fill == FALSE) {
-    bind_rows(aval_not_missing_locf, aval_missing)
-  } else {
-    aval_not_missing_locf
+  # Get the missing analysis_var (e.g., AVAL) records
+  aval_missing <- dataset %>%
+    filter(is.na(!!analysis_var))
+
+  if (imputation %in% c("add", "update_add")) {
+    bind_rows(aval_locf, aval_missing)
+  } else if (imputation == "update") {
+    aval_locf
   }
 }